requirements.txt
requests
dnspython
schedule
python-dotenv
croniter